补充docker环境依赖

+++ docker-daemon.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ -f /.dockerenv ]
+then
+		echo "this script must't run in docker!"
+		exit 1
+fi
+
 docker container prune -f
 
 docker run -d --privileged \
